KT01: DisconnectHandler.GetDisconnectToken: Fix race condition and avoid speed up read case.
This fixes a thread safety problem that the CreateToken call was not syncrhonized. Calls to GetDisconnectToken take 0.8% inclusive of WebAPI scenarios. This change speeds up the read case by avoiding the allocation of a delegate and closure and by optimizing for the more common get read case when pulling on the concurrent dictionary. It also replaces the Lazy<T> which needs 3 GC objects for the lifetime of each connection with a single GC object.